move createUser code from session to db since deleteUser and listUsers are already there
