+++ test/testsuite/tests/db.js
@@ -805,3 +805,82 @@ exports['listUsers'] = function (test)
     });
     
 };
+
+
+
+
+
+
+
+var deleteUser = function(username, test, callback) {
+    db.deleteUser(username, function(err) {
+        if (err) {
+            test.done(err);
+        } else {
+            callback();
+        }
+    });    
+};
+
+exports['createUser'] = function (test)
+{
+    test.expect(5);
+
+    async.waterfall([
+        function (callback) {
+            db.createUser('wizard_of_oz', 'password', function(err, user) {
+                if (err) {
+                    test.done(err);
+                }
+                test.equal(err, undefined, 'signup has no error');
+                test.notEqual(user, undefined, 'signup returns a value');
+                test.equal(user.ok, true, 'signup returns okay');
+                callback();
+            });
+        },
+        function (callback) {
+            db.createUser('wicked_with_of_the_east', 'password', ['witch'], function(err, user) {
+                if (err) {
+                    test.done(err);
+                }
+                
+                db.getUser('wicked_with_of_the_east', function(err, user, options) {
+                    if (err) {
+                        test.done(err);
+                    }
+                    test.equal(user.roles[0], "witch");
+                    callback();                    
+                });
+            });            
+        },
+        function (callback) {
+            db.createUser('tin_woodman', 'password', ['_admin'], function(err, user) {
+                if (err) {
+                    test.done(err);
+                }
+                
+                db.request({
+                    type: 'GET',
+                    url: '/_config/admins/tin_woodman',
+                    contentType: 'application/json'                
+                }, function(err, admin) {
+                    if (err) {
+                        test.done(err);
+                    }
+                    
+                    test.notEqual(admin, undefined, 'tin_woodman should be an admin');
+                    callback();
+                });
+            });
+        },
+        function (callback) {
+            deleteUser('wizard_of_oz', test, function() {
+                deleteUser('tin_woodman', test, function() {
+                    deleteUser('wicked_with_of_the_east', test, callback);
+                });
+            });
+        }
+    ], function () {
+        test.done();
+    });
+};
